Revert "Bump min setuptools version and remove setup.py (#154)"
This reverts commit 6eeef938c877634dcb7c8acb101bb2363ed1455b.

Numpy currently pins setuptools to 59.2 see:
https://github.com/numpy/numpy/issues/22623

This commit caused UFL to require a lated setuptools. For complex
reasons to do with blas dependencies, Firedrake builds numpy without
build isolation. We also install UFL in editable mode, so that can't
use build isolation. Consequently we are stuck on the pinned setuptools.